le dac fonctionne

+++ include/dac.h
@@ -12,7 +12,7 @@ public:
     void setOutputVoltage(double voltage) const;
     virtual void sendValue(int value) const;
 
-    static const uint8_t address = 0x60;
+    static const int address = 0x60;
     static const int minDacValue = 0;
     static const int maxDacValue = 4095;
     constexpr static double minVoltage = 0.0;
